Add a --transposed command line option and an additional "overall total" result
There are now two different ways we can process the input files: "normal" and "transposed".

"normal" means:
```
for each file
    for each parser
        parse(file)
```
whereas "transposed" means:
```
for each parser
    for each file
        parse(file)
````

I also added an extra result, called "Overall Total", which shows the sum of the parsing time on all files for each parser. This is calculated after parsing completes and gets printed as the last line in the table.

Looking at the total time for each parser in "transposed" mode should give a better idea of their relative performance for parsing large groups of PLY files, which is the case we have when loading a PBRT scene.