Use a polymesh instead of a trimesh, so that no triangulation is required
This tends to give the comparison a slightly different emphasis, since we're not having to package indices for triangulation. It simplifies the code here a little too which is nice.

This feels like it might be slightly less representative of real world use though: I pretty much always want a triangulated mesh for my own use cases. Hmm.