Add the --verts-per-face command line option
This command line option tells the parsers to assume that every file they parse has exactly N vertices per face, for every face (N = the value given on the command line). Some parsers, including miniply and tinyply, are able to take advantage of this to get a big performance increase in parsing so it's useful to be able to do perf comparisons under those conditions. Obviously it won't be a valid assumption on every file though.