Fix occasional crash when using msh_ply
The problem was that in some cases the msh_ply_t handle returned from msh_ply_open was null, which was causing a crash when I passed it to msh_ply_close.  I've moved the call to msh_ply_close inside an existing if-block that only executes when the handle is non-null & I'm flagging the case where the handle is null as a test failure.

Another problem was that I wasn't zero-initialising the property descriptors.