macOS compile fixes
The cmake line specifying that we're compiling as c++11 was accidentally removed in an earlier change. Compiling was still successful on Windows because MSVC is less strict about language versions, but was failing on macOS. Added the line back in.

msh_std.h, which is needed for the msh_ply parser, depends on some additional headers on macOS which weren't being included: <sys/stat.h> and <unistd.h>. We're no including those inside a #ifdef __APPLE__ block.