Add results on macOS including msh_ply
Interestingly, on macOS and with precognition enabled, msh_ply is quite a bit faster than miniply for two of the three collections. Yet on Windows, miniply is faster on all collections. This only applies when precognition is enabled, but it implies there are still a few tricks I can learn from msh_ply. Kudos to the author!