Flag the Overall result as failed if any of the input files failed to parse
Before this commit, the Overall time was simply the total time for all successfully parsed files. This meant that a parser which was failing to parse some (or all) files would be misleadingly reported as having a faster overall time.

With this commit the overall result for a parser is flagged as failed if any of the individual input files failed to parse. This means the summary table will now show "failed" for that parser, instead of showing a time.

This came up because tinyply was failing to parse the files from the Stanford 3D Scanning Repository after I upgraded it to the latest version, but I didn't notice because I was only looking at the summary.